Implemented setting enemy health and speed with  xml file.
