+++ data/EnemyPacks/goldroach/cockroach.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="ISO-8859-1" standalone="yes"?>
 <EnemyRoot>
   <Name>goldroach</Name>
+  <Attributes health="80" speed="2.5" /> 
   <Texture path="cockroach.png" />
   <WalkClips>
     <clip_width width="51" />
