+++ src/scripted_enemy.cpp
@@ -13,18 +13,19 @@
 #include "content_loader.h"
 
 //constructor
-ScriptedEnemy::ScriptedEnemy(std::string name, int x,int y,int width,int height) : Enemy(x,y,width,height)
+ScriptedEnemy::ScriptedEnemy(std::string name, 
+							std::int16_t health, float speed,
+							int x, int y, int width, int height) : Enemy(x,y,width,height)
 {
 	
 	m_name = name;
 	
-    //std::int16_t initialHealth = 50;
-    Enemy::setHealth(cockroachHealth);
+    Enemy::setHealth(health);
+    
     Enemy::setEnemyState(Enemy::EnemyState::MOVING_NO_PLAYER);
     ScriptedEnemy::pushState(ScriptedEnemy::ScriptedEnemyState::DETERMINE_DIRECTION);
     
     //initialize speed
-    float speed = 30;
     Enemy::setSpeed(speed);
     
     probabilitiesDirection = {0.2, 0.2,0.2,0.2,0.3};
@@ -56,260 +57,6 @@ void ScriptedEnemy::setupScriptedEnemyCollisionObject()
     Enemy::setOwnerTypeOfCollisionObject(type);
 }
 
-bool loadScriptedEnemyVisualMedia(std::string xml_file_path,std::string xml_file_dir,
-						LTexture* cTexture,
-                        std::vector <SDL_Rect> &walk_clips,
-                        SDL_Renderer* gRenderer )
-{
-    
-    // Create empty XML document within memory
-	pugi::xml_document doc;
-
-	// Load XML file into memory
-	// Remark: to fully read declaration entries you have to specify
-	// "pugi::parse_declaration"
-	pugi::xml_parse_result result = doc.load_file(xml_file_path.c_str(),
-											pugi::parse_default);
-	if (!result)
-	{
-		std::cout << "Parse error: " << result.description() << ", character pos= " << result.offset;
-		return false;
-	}
-    
-    pugi::xml_node root = doc.child("EnemyRoot");
-    
-    std::string cTexFilePath = xml_file_dir + "/" + root.child("Texture").attribute("path").value();
-    
-    //initialize texture
-    if(!cTexture->loadFromFile(cTexFilePath.c_str(),gRenderer) )
-    {
-        std::cout << "scripted enemy image loading failed! \n";
-        std::cout << "filepath:" << cTexFilePath << std::endl;
-        return false;
-    }
-    else
-    {
-		
-		std::string valString;
-		
-		//set size of walk clips vector
-		valString = root.child("WalkClips").child("clip_num").attribute("number").value();
-		size_t clipsNum = atoi(valString.c_str());;
-        walk_clips.resize(clipsNum);
-		
-		//set width and height of each uniform clips
-		valString = root.child("WalkClips").child("clip_width").attribute("width").value();
-        std::int8_t width = atoi(valString.c_str());
-        valString = root.child("WalkClips").child("clip_height").attribute("height").value();
-        std::int8_t height = atoi(valString.c_str());
-        
-        SDL_Rect clip;
-        clip.w = width;
-        clip.h = height;
-        
-        valString = root.child("WalkClips").child("UP_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_1] = clip;
-        
-        valString = root.child("WalkClips").child("UP_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_2] = clip; 
-        
-        valString = root.child("WalkClips").child("UP_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_3] = clip; 
-        
-        valString = root.child("WalkClips").child("UP_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_4] = clip;
-        
-        valString = root.child("WalkClips").child("UP_LEFT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_LEFT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_LEFT_1] = clip;
-        
-        valString = root.child("WalkClips").child("UP_LEFT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_LEFT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_LEFT_2] = clip;
-        
-        valString = root.child("WalkClips").child("UP_LEFT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_LEFT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_LEFT_3] = clip;
-        
-        valString = root.child("WalkClips").child("UP_LEFT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_LEFT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_LEFT_4] = clip;
-        
-        valString = root.child("WalkClips").child("LEFT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("LEFT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::LEFT_1] = clip;
-        
-        valString = root.child("WalkClips").child("LEFT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("LEFT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::LEFT_2] = clip;
-        
-        valString = root.child("WalkClips").child("LEFT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("LEFT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::LEFT_3] = clip;
-        
-        valString = root.child("WalkClips").child("LEFT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("LEFT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::LEFT_4] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_LEFT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_LEFT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_LEFT_1] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_LEFT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_LEFT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_LEFT_2] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_LEFT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_LEFT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_LEFT_3] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_LEFT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_LEFT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_LEFT_4] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_1] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_2] = clip;
-        
-        valString = root.child("WalkClips").child("UP_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_3] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_4] = clip;
-        
-        valString = root.child("WalkClips").child("RIGHT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("RIGHT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::RIGHT_1] = clip;
-        
-        valString = root.child("WalkClips").child("RIGHT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("RIGHT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::RIGHT_2] = clip;
-        
-        valString = root.child("WalkClips").child("RIGHT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("RIGHT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::RIGHT_3] = clip;
-        
-        valString = root.child("WalkClips").child("RIGHT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("RIGHT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::RIGHT_4] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_RIGHT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_RIGHT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_RIGHT_1] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_RIGHT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_RIGHT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_RIGHT_2] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_RIGHT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_RIGHT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_RIGHT_3] = clip;
-        
-        valString = root.child("WalkClips").child("DOWN_RIGHT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("DOWN_RIGHT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::DOWN_RIGHT_4] = clip;
-        
-        valString = root.child("WalkClips").child("UP_RIGHT_1").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_RIGHT_1").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_RIGHT_1] = clip;
-        
-        valString = root.child("WalkClips").child("UP_RIGHT_2").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_RIGHT_2").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_RIGHT_2] = clip;
-        
-        valString = root.child("WalkClips").child("UP_RIGHT_3").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_RIGHT_3").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_RIGHT_3] = clip;
-        
-        valString = root.child("WalkClips").child("UP_RIGHT_4").attribute("x").value();
-        clip.x = atoi(valString.c_str());;
-        valString = root.child("WalkClips").child("UP_RIGHT_4").attribute("y").value();
-        clip.y = atoi(valString.c_str());;
-        walk_clips[Sprite::UP_RIGHT_4] = clip;
-    }
-    
-    return true;
-}
-
-void freeScriptedEnemyVisualMedia(LTexture* cTexture)
-{
-    if(cTexture != nullptr)
-    {
-        cTexture = nullptr;
-    }
-}
 
 
 //function to load media for sprite
